ml-inference-service/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
